[INFRA] split into more header files
