+++ .gitignore
@@ -29,3 +29,6 @@
 # Executables
 *.exe
 *.app
+
+# Build directory
+/build/
