+++ include/raptor/search/minimiser_model.hpp
